+++ src/chains/Base.sol
@@ -33,6 +33,7 @@ import {aerodromeFactory, aerodromeInitHash, aerodromeForkId} from "../core/univ
 import {alienBaseV3Factory, alienBaseV3ForkId} from "../core/univ3forks/AlienBaseV3.sol";
 import {baseXFactory, baseXForkId} from "../core/univ3forks/BaseX.sol";
 import {swapBasedV3Factory, swapBasedV3ForkId} from "../core/univ3forks/SwapBasedV3.sol";
+import {dackieSwapV3BaseFactory, dackieSwapV3ForkId} from "../core/univ3forks/DackieSwapV3.sol";
 
 // Solidity inheritance is stupid
 import {SettlerAbstract} from "../SettlerAbstract.sol";
@@ -122,6 +123,10 @@ abstract contract BaseMixin is FreeMemory, SettlerBase, MaverickV2 {
                     callbackSelector = uint32(IUniswapV3Callback.uniswapV3SwapCallback.selector);
                 } else if (forkId == swapBasedV3ForkId) {
                     factory = swapBasedV3Factory;
+                    initHash = pancakeSwapV3InitHash;
+                    callbackSelector = uint32(IPancakeSwapV3Callback.pancakeV3SwapCallback.selector);
+                } else if (forkId == dackieSwapV3ForkId) {
+                    factory = dackieSwapV3BaseFactory;
                     initHash = pancakeSwapV3InitHash;
                     callbackSelector = uint32(IPancakeSwapV3Callback.pancakeV3SwapCallback.selector);
                 } else {