+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Master list of UniV3 forks:
   11. Fenix (Algebra-like)
   12. ZebraV3
   13. Lynex (Algebra-like)
+  14. DackieSwapV3
 
 ---
 
@@ -27,6 +28,7 @@ Master list of UniV3 forks:
 * Add MaverickV2 to Bnb
 * Add MaverickV2 to Scroll
 * Add MaverickV2 to Sepolia
+* Add DackieSwapV3 UniV3 fork to Base
 
 ## 2024-07-29